+++ packages/button-group/src/index.d.ts
@@ -1,6 +1,0 @@
-import { Component } from 'react';
-
-interface Props {}
-interface State {}
-
-export default class extends Component<Props, State> {}