+++ packages/spotlight/src/index.js
@@ -1,9 +1,0 @@
-export {
-  Spotlight,
-  SpotlightManager,
-  SpotlightTarget,
-} from './components';
-
-export {
-  Pulse,
-} from './styled/Target';