+++ packages/xflow/src/index.jsx
@@ -1,5 +1,5 @@
 export { default as RequestOrStartTrial } from './request-or-start-trial';
-export { default as RequestProductTrialOptOut } from './request-trial-opt-out';
+export { default as XFlowOptOut } from './xflow-opt-out';
 
 export { default as JiraToConfluenceXFlowProvider } from './product-xflow-providers/JiraToConfluenceXFlowProvider';
 export { default as JiraToJCXFlowProvider } from './product-xflow-providers/JiraToJCXFlowProvider';
@@ -9,3 +9,5 @@ export { default as JiraToJSWXFlowProvider } from './product-xflow-providers/Jir
 // To allow consumers not using @atlaskit/analytics
 // to delegate the analytics events to their own stack
 export { default as XFlowAnalyticsListener } from './common/components/XFlowAnalyticsListener';
+
+export { default as XFlow, UnknownProductError } from './XFlow';