+++ packages/xflow/src/request-or-start-trial/components/ConfirmTrial.jsx
@@ -8,6 +8,7 @@ import { withAnalytics } from '@atlaskit/analytics';
 import ErrorFlag from '../../common/components/ErrorFlag';
 
 import SpinnerDiv from '../../common/styled/SpinnerDiv';
+import OptOutLinkButton from '../styled/OptOutLinkButton';
 import StartTrialFooter from '../styled/StartTrialFooter';
 import StartTrialHeader from '../styled/StartTrialHeader';
 import { withXFlowProvider } from '../../common/components/XFlowProvider';
@@ -40,6 +41,7 @@ class ConfirmTrial extends Component {
     cancelStartProductTrial: PropTypes.func,
     onComplete: PropTypes.func.isRequired,
     onCancel: PropTypes.func.isRequired,
+    isCrossSell: PropTypes.bool,
 
     intl: intlShape.isRequired,
     firePrivateAnalyticsEvent: PropTypes.func.isRequired,
@@ -110,6 +112,21 @@ class ConfirmTrial extends Component {
       .then(onCancel);
   };
 
+  handleOptOutClick = () => {
+    const { firePrivateAnalyticsEvent, status } = this.props;
+    firePrivateAnalyticsEvent(
+      status === INACTIVE
+        ? 'xflow.confirm-trial.opt-out-button.clicked'
+        : 'xflow.reactivate-trial.opt-out-button.clicked'
+    );
+    this.setState({
+      spinnerActive: true,
+      buttonsDisabled: true,
+    });
+    // Redirect to GAB, with opt-out parameter so the opt-out dialog opens
+    window.location.href = '/admin/billing/addapplication?requestproductoptout=true';
+  };
+
   handleErrorFlagDismiss = () => {
     const { firePrivateAnalyticsEvent, status } = this.props;
     firePrivateAnalyticsEvent(
@@ -131,6 +148,7 @@ class ConfirmTrial extends Component {
       trialMessage,
       reactivateHeading,
       reactivateMessage,
+      isCrossSell,
     } = this.props;
     return (
       <ModalDialog
@@ -173,6 +191,14 @@ class ConfirmTrial extends Component {
           </StartTrialHeader>
           {status === INACTIVE ? trialMessage : reactivateMessage}
         </div>
+        {isCrossSell === true &&
+          <OptOutLinkButton
+            id="xflow-opt-out-button"
+            onClick={this.handleOptOutClick}
+          >
+            Turn off these messages
+          </OptOutLinkButton>
+        }
         <ErrorFlag
           title={intl.formatMessage(messages.errorFlagTitle)}
           description={intl.formatMessage(messages.errorFlagDescription)}