+++ packages/xflow/src/request-or-start-trial/components/ContextualStartTrial.jsx
@@ -22,6 +22,7 @@ class ContextualStartTrial extends Component {
 
     onComplete: PropTypes.func,
     onTrialActivating: PropTypes.func,
+    isCrossSell: PropTypes.bool,
 
     firePrivateAnalyticsEvent: PropTypes.func.isRequired,
   };
@@ -29,6 +30,7 @@ class ContextualStartTrial extends Component {
   static defaultProps = {
     onComplete: () => {},
     onTrialActivating: () => {},
+    isCrossSell: false,
   };
 
   render() {
@@ -38,6 +40,7 @@ class ContextualStartTrial extends Component {
       onTrialActivating,
       firePrivateAnalyticsEvent,
       contextInfo,
+      isCrossSell,
     } = this.props;
     return (
       <MultiStep start={0} onComplete={onComplete}>
@@ -55,6 +58,7 @@ class ContextualStartTrial extends Component {
               }}
               onCancel={cancel}
               contextInfo={contextInfo}
+              isCrossSell={isCrossSell}
             />}
         />
         <Step render={nextStep => <GrantAccess onComplete={nextStep} />} />