+++ packages/xflow/src/xflow-opt-out/styled/OptOutIcon.jsx
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+const OptOutIcon = styled.span`
+  padding: 0px 8px 0px 0px;
+`;
+
+OptOutIcon.displayName = 'OptOutIcon';
+export default OptOutIcon;