+++ packages/xflow/stories/request-or-start-trial/components/ConfirmTrial-story.jsx
@@ -75,4 +75,10 @@ storiesOf('request-or-start-trial/ConfirmTrial', module)
       status={DEACTIVATED}
       startProductTrial={() => new Promise((_, reject) => setTimeout(reject, 1500))}
     />
+  )
+  .add('Confirm Trial dialog (INACTIVE) with opt out link', () =>
+    <ConfirmTrialBase
+      {...defaultProps}
+      isCrossSell
+    />
   );