+++ packages/task-decision/docs/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @atlaskit/task-decision
 
+## 4.7.2 (2018-01-29)
+
+* bug fix; object/containerAri is optional to toggle isDone for ResourcedTaskItem (issues closed: fs-1674) ([1f77a21](https://bitbucket.org/atlassian/atlaskit/commits/1f77a21))
+## 4.7.1 (2018-01-18)
+
+* bug fix; move getCurrentUser to TaskDecisionResource ([222ef0e](https://bitbucket.org/atlassian/atlaskit/commits/222ef0e))
+## 4.7.0 (2018-01-16)
+
+
+
+* feature; product can pass username to resource to prevent delayed completed by message (issues closed: fs-1539) ([b85c8ac](https://bitbucket.org/atlassian/atlaskit/commits/b85c8ac))
+
+
 ## 4.6.1 (2017-12-19)
 
 * bug fix; bump packages to fixed version of analytics ([615e41c](https://bitbucket.org/atlassian/atlaskit/commits/615e41c))